refactor: extract shared chat composer with connectors, paste, and drag-drop
- Extract shared ZeroChatComposer component from duplicated code in talk
  and session chat pages
- Move connector logic (popover, add dialog, connect modal) into the
  shared composer so both pages get identical connector functionality
- Fix drag-drop highlight not showing (outline instead of ring to avoid
  box-shadow conflict with .zero-composer CSS)
- Fix paste uploading file name instead of file content
- Make AttachmentChip non-exported (internal to AttachmentChips)
